cleaner and prettier directories
