+++ src/main/resources/static/style.css
@@ -163,3 +163,61 @@ video {
   top: 2rem;
   align-self: flex-start;
 }
+
+/* Pinned launcher bar */
+#pinned-launchers {
+  position: sticky;
+  top: 0;
+  background: white;
+  z-index: 10;
+  padding: 1rem 0;
+  margin-bottom: 1rem;
+  border-bottom: 2px solid #ddd;
+}
+
+/* Keep button layout clean inside pinned */
+#pinned-launchers ul {
+  display: flex;
+  gap: 1rem;
+  flex-wrap: wrap;
+}
+
+/* Launcher buttons */
+#pinned-launchers button {
+  flex: 1;
+  min-width: 150px;
+  padding: 0.6rem;
+  font-weight: bold;
+  border-radius: 8px;
+  background: #f7f7f7;
+  border: 2px solid #ccc;
+  box-shadow: 1px 1px 3px rgba(0,0,0,0.1);
+  transition: 0.2s all;
+}
+
+#pinned-launchers button:hover:not(.disabled) {
+  background: #e0e0e0;
+  transform: scale(1.02);
+}
+
+/* Grayed out look when loading */
+.media-btn.disabled {
+  opacity: 0.5;
+  pointer-events: none;
+}
+
+#loading-spinner {
+  display: none;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  font-size: 1.5rem;
+  font-weight: bold;
+  background: rgba(0,0,0,0.8);
+  color: white;
+  padding: 1rem 2rem;
+  border-radius: 8px;
+  z-index: 9999;
+}
+
